rename interface file to what seems to be the standard ? IProfInfo...
